refactor: move PurchaseAsAnonymousConsequences components
