--- src/anonymous-purchase-consequences-screen/index.ts
@@ -1,1 +1,0 @@
-export {AnonymousPurchaseConsequencesScreenComponent} from './AnonymousPurchaseConsequencesScreenComponent';
